move add_ingots tests to own file
