+++ tests/main_test.py
@@ -49,83 +49,6 @@ class TestIronForgedBot(unittest.IsolatedAsyncioTestCase):
         self.mock_interaction.guild.members = []
 
     @patch("main.validate_protected_request")
-    async def test_addingots(self, mock_validate_protected_request):
-        """Test that ingots can be added to a user."""
-        leader_name = "leader"
-        playername = "johnnycache"
-        player_id = 123456
-
-        mock_validate_protected_request.return_value = (
-            helper_create_member(leader_name, ROLES.LEADERSHIP),
-            playername,
-        )
-
-        mock_storage = MagicMock()
-        mock_storage.read_member.return_value = Member(
-            id=player_id, runescape_name=playername, ingots=5000
-        )
-
-        bot = main.IronForgedCommands(MagicMock(), MagicMock(), mock_storage, "")
-
-        await bot.addingots(self.mock_interaction, playername, 5000)
-
-        mock_storage.update_members.assert_called_once_with(
-            [Member(id=player_id, runescape_name=playername, ingots=10000)],
-            leader_name,
-            note="None",
-        )
-
-        self.mock_interaction.followup.send.assert_called_once_with(
-            f"Added 5,000 ingots to {playername}; reason: None. They now have 10,000 ingots "
-        )
-
-    @patch("main.send_error_response")
-    @patch("main.validate_protected_request")
-    async def test_addingots_player_not_found(
-        self, mock_validate_protected_request, mock_send_error_response
-    ):
-        """Test that a missing player is surfaced to caller."""
-        leader_name = "leader"
-        playername = "johnnycache"
-
-        mock_validate_protected_request.return_value = (
-            helper_create_member(leader_name, ROLES.LEADERSHIP),
-            playername,
-        )
-
-        mock_storage = MagicMock()
-        mock_storage.read_member.return_value = None
-
-        commands = main.IronForgedCommands(MagicMock(), MagicMock(), mock_storage, "")
-        await commands.addingots(self.mock_interaction, playername, 5)
-
-        mock_send_error_response.assert_awaited_with(
-            self.mock_interaction, f"Member '{playername}' not found in spreadsheet"
-        )
-
-    @patch("main.send_error_response")
-    async def test_addingots_permission_denied(
-        self,
-        mock_send_error_response,
-    ):
-        """Test that non-leadership role can't add ingots."""
-        guild = AsyncMock(discord.Guild)
-        caller = helper_create_member("1eader", ROLES.MEMBER)
-        member = helper_create_member("member", ROLES.MEMBER)
-        guild.members = [caller, member]
-        self.mock_interaction.user = caller
-        self.mock_interaction.guild = guild
-
-        commands = main.IronForgedCommands(MagicMock(), MagicMock(), MagicMock(), "")
-
-        await commands.addingots(self.mock_interaction, member.name, 5)
-
-        mock_send_error_response.assert_awaited_with(
-            self.mock_interaction,
-            f"Member '{caller.display_name}' does not have permission for this action",
-        )
-
-    @patch("main.validate_protected_request")
     def test_addingots_bulk(self, mock_validate_protected_request):
         """Test that ingots can be added to multiple users."""
         leader_name = "leader"
